simple background, shop foundation, game-over mechanic

+++ src/main/java/businessdirt/dodgecoin/core/config/Config.java
@@ -17,7 +17,7 @@ public class Config extends ConfigHandler {
             category = "Hidden",
             description = "Money"
     )
-    public static int money = 0;
+    public static float money = 0;
 
     public Config() {
         super(new File(Util.getConfigFolder(), "\\config.toml"));
